requirements.txt
python==3.9
pygame
librosa
opencv-contrib-python
mediapipe
pytorch